seperate views file
