+++ blog/templates/blog/question_detail.html
@@ -69,7 +69,9 @@
         <div class="answer-box">
             <div class="answer-status">
                 {% if comment.id_code == blog.solution_id_code %}
-                    <div class="solution-button" style="background-color: darkgreen;" onclick="location.href=`{% url 'blog:choose-solution' comment.id_code %}`"></div>
+                    <div class="solution-button" style="background-color: darkgreen;" onclick="location.href=`{% url 'blog:choose-solution' comment.id_code %}`">
+                        <p style="font-size: xx-large;margin-left: 25%;">&#10003;</p>
+                    </div>
                 {% else %}
                     <div class="solution-button" onclick="location.href=`{% url 'blog:choose-solution' comment.id_code %}`"></div>
                 {% endif %}
