+++ blog/urls.py
@@ -1,21 +1,21 @@
 from django.urls import path
 
-from . import views
+from .views import *
 
 app_name = 'blog'
 urlpatterns = [
-    path('', views.blog, name='blog-index'),
-    path('create-blog/<blog_type>', views.create_blog, name='create-blog'),
-    path('<blog_type>', views.filter_blog, name='filter-blog'),
-    path('<id_code>/', views.blog_detail, name='blog-detail'),
-    path('<post_id_code>/edit-blog/', views.edit_blog, name='edit-blog'),
-    path('<post_id_code>/delete-blog/', views.delete_blog, name='delete-blog'),
-    path('<post_id_code>/crete-comment', views.create_comment, name='create-comment'),
-    path('<comment_id_code>/delete-comment', views.delete_comment, name='delete-comment'),
-    path('<comment_id_code>/create-subcomment', views.create_subcomment, name='create-subcomment'),
-    path('<comment_id_code>/choose-solution', views.choose_solution, name='choose-solution'),
-    path('tag/<tag_name>', views.tag, name='tag'),
-    path('poll/vote/<choice_id_code>',views.vote, name='vote'),
-    path('like/<id>',views.like, name='like'),
-    path('apply-job/<job_id_code>',views.apply_job, name='apply-job'),
+    path('', blog, name='blog-index'),
+    path('create-blog/<blog_type>', create_blog, name='create-blog'),
+    path('<blog_type>', filter_blog, name='filter-blog'),
+    path('<id_code>/', blog_detail, name='blog-detail'),
+    path('<post_id_code>/edit-blog/', edit_blog, name='edit-blog'),
+    path('<post_id_code>/delete-blog/', delete_blog, name='delete-blog'),
+    path('<post_id_code>/crete-comment', create_comment, name='create-comment'),
+    path('<comment_id_code>/delete-comment', delete_comment, name='delete-comment'),
+    path('<comment_id_code>/create-subcomment', create_subcomment, name='create-subcomment'),
+    path('<comment_id_code>/choose-solution', choose_solution, name='choose-solution'),
+    path('tag/<tag_name>', tag, name='tag'),
+    path('poll/vote/<choice_id_code>',vote, name='vote'),
+    path('like/<id>',like, name='like'),
+    path('apply-job/<job_id_code>',apply_job, name='apply-job'),
 ]
