+++ blog/views/__init__.py
@@ -1,0 +1,8 @@
+from .blog_views import *
+from .post_views import *
+from .question_views import *
+from .poll_views import *
+from .job_views import *
+
+from .comment_views import *
+from .tag_views import *
